Moved GetAliveNodes and isLeader from cluster to mdnsdiscover
